Restructure cmake files
Move the C applications to the applications directory (from
src/applications) and the build- and install instructions to
application/'s cmake file.

The language binding cmake files now internally handle the logic and
variables as to if they're to be compiled or not, meaning the root cmake
only has to include the python/mex directory.

Warning compiler flag logic has been moved to cmake/default_warnings
which will handle compiler differences, so that too much platform
dependent noise isn't in the root cmake file.

Some absolute paths are used over relative paths since it's no longer a
guarantee that files in the cmake directory will be included or target
something directly from the CMAKE_SOURCE_DIR.