Primitive mmap support
If the system has the mmap system call available (really any posix
compliant system) the segy_mmap function will attempt to memory map the
file in question.

This is considered an experimental, optional feature and the scheme is
rather naïve - the full file will be mmap'd, not just a view into it,
meaning it will likely cause forms of system failure for
way-larger-than-memory files. However, with non-regular access such as
reading lines or every nth trace or similar performance is vastly
improved.