segy_count_lines, new interface for line count
The segy_lines_count makes the minimum set of assumptions, at the expense of
usability. Add the segy_count_lines function which has a simpler interface
without relative argument positioning.

In most cases this should be the preferred function because you usually *do*
have all the pieces of information you need to use this simpler function.