Namespace enums.
Prefixes the enums with SEGY_, so that they no longer introduce
potential name conflicts with common local variables. All enums and
other global constants are now underscore-separated all-caps
identifiers.

A consequence of this namespacing is that identifiers become *very* long
and verbose, but on the flipside it makes it rather clear when something
is a aliased constant. The worst offender is offset (field 37 in trace
headers) which is a commonly used name for local variables.