Make file.samples a list, not scalar
Analoguous to file.xlines, file.ilines, file.samples is now a list of
samples and their identifiers. If possible, this is a list derived from
the binary- and trace headers with a recording delay t0 and a sample
interval. If these can't be figured out they're assumed to be 0 and
4000μs (4ms) respectively.

As with the other identifiers, the cardinality is now obtained with
len().