Don't assume time for sample_interval
The segy_sample_interval function would always assume that the interval
between sample describes distance-in-time (ms), but this means
information is lost was in meters (which would be interpreted as
millimeters). Instead, this function now returns the inferred interval
without modifications, leaving it up to the caller to determine what the
interval describes, and typically has more information available.

This means that the sample_indexes function in python.tools now changes
behaviour and uses microseconds (as per the specification) rather than
milliseconds. To support the typical use of this function,
segyio.tools.scale_samples has been added.