Do not throw on files without good geometry
Support opening files without a reasonable geometry, or that are in some
other way unsorted. Opens up a use case where segyio is used to read
files not yet stacked, organised by lines, but still has useful
properties or interesting data.

The default behaviour is to assume a file is sorted, and raise an
exception if this cannot be figured out.