Remove scale_samples; store samples as float
The heuristic that separates between time/depth is broken, since the
sample rate for depth is millimeters, not meters. In other words, it is
impossible to figure out just by inspecting the file if sample interval
means time or depth. The entire tool is removed in favour of correctly
storing the samples as floats.