Don't free caller provided buffers in subtr
The read/write subtr functions can take a buffer for when the function
is to be called repeatedly for speed purposes, but will allocate
otherwise. However, the functions always tried to free this buffer, even
if it wasn't allocated by the function.

Check if the buffer was actually allocated by the functions before
attempting to free.

Tests are updated to check for this.