Upload source as segyio-$ver.tar.gz
The name segyio-v$ver-all.tar.gz was picked to signal that it included
submodules, but since it's largely an implementation detail and the
focus is just "buildable distribution", the name is changed.

Free software tarballs are often released as $pkg-$ver (meaning distro
tools tend to work nicely with that pattern), so travis uploads with
that name as well.