Buffer gather-ranges by reading lines
Doing ranges over gather was very slow, because it's essentially a
trace-by-trace read with the additional cost of a line-number
intersection -> trace index translation. This work was done in pure
python and once-per-trace with pretty terrible performance.

The immediate use case for gather mode is obviously looking up just a
handful of traces, but the examples and obvious code also considers
ranges, which translate to sub cubes.

In order to improve the performance of the range-over-gathers feature, a
simple way is to simply read more traces as once an buffer them, and
avoid doing too many and too expensive index translations.

Since range-over-gather essentially boils down to reading along an
inline, using the inline mode to do the physical reading, this can be
used to buffer traces. Crosslines have to be translated from numbers ->
index, but this no longer involves multiple function calls.