Build python library with setuptools
Instead of trying to generate a setup.py target from cmake variables
that's robust, leverage native setup.py and integrate it as a step in
the regular make build process.

Using setuptools to build the extension means a lot of the complicated
link magic, python flags etc. goes away, and creates a clearer
separation between the two components. The python library can now also
be built without the libsegyio C core, assuming it's available to link
from elsewhere.

For end users, this shouldn't mean much as they either use pre-packaged
installs or cmake && make, which works as before. It does give
maintainers more flexibility.

The currently lost feature is to build the full library with static
linking from source on pypi, as the libsegyio core isn't distributed
that way currently.

The python package use scm_versioning now, which means that it might
slightly deviate in post-patch naming from the libsegyio core -
however, they don't rely on identical version numbers.

The CI specifications are brought up-to-date with this change.