Read version from git
Read the project version from git when building. If it's a tarball
build, assume python/segyio/version.py is written (by travis, who makes
the tarballs) and read from that. Fallback to 0.0.0.

Any version number is a cache variable and can be overridden by command
line arguments.

With this change, the git tag is the authority on version, and
everything should respect that.