Verify slice step is never zero in forall
Define the slicelength helper function to be zero if the step is zero,
which avoids a code path that would trigger division-by-zero. This was
always a part of the contract to the for-all functions, which is never
invoked with a step size of 0.

This is a hint to optimisers and static analysers.