Allow creating unstructured files, relax spec
The create function would reject any spec that didn't contain
iline/xline/offset fields (created by segyio.spec() and extended),
effectively prohibiting the creation of unstructured files (i.e. a bunch
of traces).

The docs are improved to describe exactly what is expected by the spec
argument to create() to make a new file with certain properties,
including the formalisation of how to obtain unstructured files. An
unstructured file is assumed whenever either of the sorting, line or
offset fields are unset, None, or empty lists.