segyio-catr: exit with exit() instead of return
When aborting, either becuse of --help, --version, invalid input
argument or file errors, using return makes some static analysers break
on lifetime and leak analysis.

While they're right in the sense that it is a function that would leak,
it would immediately lead to process teardown.

However, exit is probably clearer than return and makes the issue go
away, without trying to silence the errors or pollute the code with too
much resource handling (at which point C++ is the obvious choice
anyway).