Omit lib/test/segy.c from clang static analysis
The test/segy.c file trigger static analysis warnings such as
uninitialised values (that are initialised in functions), failing the
build on travis, but such warnings from here aren't very interesting
(especially if ctest pass).

They're omitted for now until a better test suite is in place or the
values are initialised properly and only actual problems are reported.