Help delocate find dylibs on OS X on travis
delocate is used to build wheels on travis, and the logic that looks up
run-time dependencies could not figure out where the correct .dylib was
from @rpath/libsegyio.1.dylib. This is apparently expected behaviour.
From delocate's readme:

> where the first line is the install name id that the linker picked up
> when linking myext.so to libme.dylib. You job is to fix the build
> process so that libme.dylib has install name id /path/to/libme.dylib.
> This is not a problem specific to delocate; you will need to do this
> to make sure that myext.so can load libme.dylib without libme.dylib
> being in the current working directory. For CMAKE builds you may want
> to check out CMAKE_INSTALL_NAME_DIR.

https://pypi.python.org/pypi/delocate

By passing CMAKE_INSTALL_NAME_DIR to cmake (which does nothing on linux
as far as I'm aware), multibuild is able to pick up the correct dylib
and build it into the wheel.