Move interpret_file test to catch2
Move the structure inferrence tests to the catch2 framework, deprecating
the old tests in C. This introduces a dependency on a C++11 compliant
compiler to run the tests. In return, the tests are annoteded and
clearer, more robust and easier to extend, and won't stop execution on
first failure.

The old C tests are renamed to c.old.$name, and the c.$name tests are
all catch based.