Replace capsule with python class in extension
With the move to C++98 for the python extension, plenty things become
more practical to do. As it turns out, making a proper python class
solves a lot of issues that plagues the pycapsule("FILE*") solution,
such as checking that the passed self object is of right type.

The move is a direct port to C++ features, with the addition of a
smarter file handle that, when a file is closed, will on subsequent
pointer-gets set the approperiate exception, that mimics python's own
file object.