Copy dll into source tree when building on windows
There is no auditwheel-like tool for python wheels on windows, which
means segyio must manually bundle the segyio core dll with the wheel.
setuptools is *very* pick and assumes everything is relative paths
downwards, and always run from root in the source directory, which means
grabbing a copy of the just-in-time built dll is non-trivial.

On windows, copy the dll into the source directory, like version.py is
written by setuptools.This ensures setuptools find the dll when it's
making the python package, and copies it in accordingly.