Allocate header-buffers from python
Remove the allocate-in-capsule logic, in favour of allocating buffers in
python and accessing them as-is with the Py_buffer interface. The bytes
object's immutability is a problem, since plenty of the internal buffers
in segyio are for writing, so bytearray is the better type. The
bytearray should rarely leak into user code.

The query-size functions are moved into the anonymous namespace,
renamed, and uses the PyLong interface, rather than BuildValue.