Read trace0/bsize/metrics during init in python
Move the metrics method into init, and store the trace0, trace_bsize and
friends inside the file handle object, along side the file pointer.
These are file specific and used for nearly every method call, and more
efficient and natural to store inside the file handle than as members in
the user-facing class. This removes the need for some of the internal
variables too, at the cost of juggling more information in C++.