Set error exit status on failure in segyio-cath
Something went wrong when the segyio-cath tests were set up, and the
application has been subtly broken for a while, as error codes were not
properly set on invalid input arguments or broken input files.

* Return non-zero on (strict) cath errors
* Return non-zero on zero input arguments

segyio-cath would return errno on no-such-file or file-seek errors when
input was non-existent or broken, but these errors typically do not set
errno, meaning it was interpreted as success.