Manage filehandle automatically in pyobject init
Use C++ constructors-destructors to manage close-on-exit, rather than
manually tracking all exit paths in init. It's basically a
std::unique_ptr without move semantics (since no C++11 yet).

The member functions swap and close were added. Close, because having to
close and remember to set internal pointer to NULL is easy to miss and
get wrong, and swap in order to easily always close either the
invalid-but-open file handle in init or the old handle in case of a
re-run init.

The autofd constructor is removed, in order to make it a POD type. It
should be more robust in combination with python C operations, being a
C-compatible struct, and the constructor was never really used anyway.