Manage all Py_buffers with buffer_guard
Extend the buffer_guard object to be a stronger automation tool on
Py_buffer invariants.

Py_buffers can be obtained both through GetBuffer calls and directly
from parsetuple, but in both cases the PyBuffer objects need to release
the buffer with PyBuffer_Release. Instead of having just a
pointer-to-buffer as member, the buffer_guard contains a full pybuffer
object.

From this, obtain-pybuffer error handling is centeralised in the
buffer_guard class, as are details on how to get it.

Since Py_buffer is a small C struct, copying it is basically free.

Overloading the operator& in buffer_guard means that the buffer_guard
works as a drop-in with python's C-API which takes pointers to the
buffers anyway.