Redesign exception messages in python extension
Rewrite all exception messages across the python extension, and reflow
code accordingly. Some simple aliases are used to make the exception
code more ergonomic and less noisy at call site.

The main focus of this patch is the improved error messages, but because
of the intrusive nature of changing to helper-functions and changing
string literals, and sometimes changing to formatted strings, some
control flow is rewritten.