Return NOTFOUND in count on unstructured files
Not being able to count lines isn't a read/write failure, but rather a
property of the input being faulty. Introduce the new SEGY_NOTFOUND
error that describes an inability to return a proper value, but not
because the file is inherently broken, just insufficiently structured.

This allows python to more accurately describe a the problem (and a
solution) when an arbitrary file is given to segyio.open, rather than a
generic failed-to-read file, which is meant for more exceptional cases.