Merge python unit tests to one pytest run
Starting pytest on a per-file basis is a lot slower than loading
individual files with unittest. By merging the tests into a single
pytest invocation (as a module, so there's no looking up paths to the
pytest binary), the total running time for the tests is virtually
unchanged, at the expense of some lack of filtering of python unit tests
from the ctest -R and -I options.

This is acceptable because pytest's own runner being a lot beefier than
unittest, and the filtering and individual test-runs is deferred to the
user by virtue of manually invoking pytest. However, little actual
utility is lost, as the report pytest generates is more comprehensible
and detailed and failure-specific than that of unittest.