Added bandit.yml file
When running pytest the preffered way of doing assertions is to use the
assert statement. Assertions is by default flagged by bandit.

To be able to configure how we want bandit to run, we have added the
bandit.yml file, and added one configuration which tells bandit to skip
assert checks. For this configuration file to be considered we have
added an arg in .travis.yml "before_script" section at - bandit ...