Get tracecount from arg in create, not header
The cheeky trick of passing the number of traceheaders in the binary
header when creating new files, in a field that is not dedicated to
number-of-traces-per-file, doesn't work because it's only 16 bits and
overflows very fast, and SEG-Y allows for 32-bits of traces (governed by
traceno).

Instead, pass the override-number-of-traces argument to the internal
filehandle constructor explicitly, and only consider that number if the
binary header is also passed.

This fixes the issue reported in
https://github.com/Statoil/segyio/issues/235