Delete wheels before attempting source upload
If wheels are present in the dist/ directory, the pypi deploy mechamism
will try to upload these as well as the source distribution, which then
fails because the name already exists on pypi.

To make the source/binary distinction more robust and elegant, use -z
and if [[ $VAR ]] checks to figure distinguish source and binary
deploys.