Relax (from f32) type restrictions on trace write
Accept any array_like and convertible-to-float to trace assignment, and
warn instead of raising an error. It allows users to assign from arrays
and double-arrays, assuming they accept the cost of conversion, and lack
of control of narrowing.

In most use cases, the narrowing would be the numpy-provided
double-to-float copy anyway, so nothing is lost except for the extra
operations required.