Lazily import the python extension
If the module is lazily imported, there's no need to mock it for
documentation purposes, and importing the source library works in the
source tree (for stuff like version).

This will push errors related to the extension misbehaving or not being
built until runtime, but tests should catch that rather fast.