Implement Field class with collections.mapping
The collections.MutableMapping is the upstream base interface for a
dict-like structure, which the Field class tries to implement, and helps
ensure that the Field types behave as predictably as possible.

Some concrete implementations disappear in this effort, rather
inheriting them from collections. Additionally, the semantics of
__getitem__ and update are cleaned up, and update has received another
guarantee, making it properly exception safe.

Docstrings are added in this pass, as is a test for the atomicity of
update.