Don't take buffer argument in Field constructors
Prefer the construction of buffers underpinning the Field class in
initialisation, instead of managing from the outside. It's a step in the
process of streamlining and formalising the Field behaviour in terms of
interface, instead of accidental behaviour through direct
buffer-manipulation in the Header implementation.

This does mean some more buffer reallocations currently, but that hit is
expected to go away soon.

The new reload() function unifies and more robustly handles the IOErrors
raised on creating headers in new files.