Explicit Trace.__iter__ over Sequence.__iter__
The default implementation of Sequence.__iter__ is nice, but assumes
that the cost of individual __getitem__ is reasonably low. In the case
of segyio, an optimisation kicks in on when querying ranges of traces,
because we know buffers can be reused.

`for trace in f.trace:` is considered idiomatic in segyio, so this
should be fast. I did some measurements when __iter__ used the
Sequence.__iter__ implementation, with this simple test script:

with segyio.open('small.sgy') as f:
    for trace in f.trace: pass
    [10000] repeats

Not doing anything in the loop body ensured only the loop cost was
measured. With Sequence.__iter__, from time:
real    0m1.470s
user    0m1.424s
sys     0m0.292s

Whereas doing `for trace in f.trace[:]: pass` 10000 times:
real    0m0.862s
user    0m0.856s
sys     0m0.256s

cProfile showed that the difference is largely numpy.zeros, the
allocator for the buffer. This patched removes this difference, by
essentially making iter(f.trace) an alias to f.trace[:].