Simplify Line class, add custom Line in header
Rewrite and simplify the Line class. Because the f.header.line mode was
implemented through this, it's also written, but still relies on the
main Line class for most details (it customises __getitem__ and
__setitem__).

Writing lines is now done by the native extension, which removes some
complexity from the python code, and provides a nice speed boost.

Proper docstrings are added for the newly rewritten objects.