Move array-conversion to function utils.castarray
In order to be more robust on different input types, writing accepts
array_likes, not just raw numpy arrays. The ideal is for the caller to
provide the right type in the first place (which is faster and safer),
so warning is approperiate. Since this procedure is the same regardless
of source and shape, it makes sense to use the same implementation in
all code paths.