Document trace in class, not in the property
In order to make the documentation more accessible, navigatable, and
concentrated, move the docstrings closer to the implementation in the
Trace class, and on the methods that provide the feature.

Originally, the intention was to make the Trace class private and
inaccessible for users, which it to some extent still is, but in order
to simplify lots of code and just remove unnecessary hurdles it's been
refactored to provide a richer experience. It doesn't make sense at all
outside the context of the file handle (which it depends on internally),
but there isn't too much wrong about clients constructing their own
accessors or even mocking in the future.

Moving the docstrings to the module aligns with this transformation.