Suppress cppcheck false positive in slicelength
cppcheck 1.84 introduces a false positive sourced in the value-flow
analysis not picking up on the division-by-zero guard when step is zero.

There is a bug report submitted upstream [1], but until it's fixed this
must be suppressed, or a cppcheck <1.84 must be used.

Pinning the cppcheck version on homebrew is clumsy at best, and a
suppression is less likely to create problems for other users.

[1] https://trac.cppcheck.net/ticket/8482