Fix an issue where segyio-catr printed the same trace twice
While segycatr takes range arguments as 1-indexed, it would still
print the first trace if 0 was given as --range or -trace. Hence, situasions
where the first trace was printed twice would occur. Now, zero is no longer
accepted as a vaule for --range or --trace. This behavior is added to man-page