Move write-subtrace tests to test-cases
Move the test-write-subtrace family of tests to test-cases, rather than
clumping them in a scenario.

The fixture for these test cases is somewhat elaborate, but does ensure
that the relationship between input-range and output vector is made
clearer. It ends up also automating the somewhat tedious checking of
whether or not a write is observable, at the cost of distancing that
check from the write-code.