Files w/ unchanged tuple (il,xl,off) are unsorted
A file is considered sorted as long as exactly one of the values
(il, xl, off) is incremented from one trace to the next, where off
should be the least significant.

il, xl, off
(0, 0, 	0)
(0, 0, 	1)
(0, 0,	2)
(0, 1, 	0)

This recognizes files where (il, xl, off) are non-changing as unsorted
rather than sorted, which was the case previously.