Make segyio.su a proper module
segyio.su is a proper module, with a directory and an __init__.

This change is motivated by a dedicated seismic unix file handle, and
genera support for the data format. The su namespace exports an
enumerator segyio.su - this is preserved through __init__ re-export and
should not break even with later, new segyio.su namespace features.