Get cmake3.x from pip in manylinux image
Instead of downloading a pre-built binary in a tarball, get cmake from
the pip infrastructure. This is a manylinux image, built in the same
image as the segyio binary package.

While code-wise this is a simple change, it's a step towards bumping the
minimum require cmake version from 2.8.12 to the 3.x series.