Use two-phase init for python file handle
Split the python file handle initialisation routine into two phases -
one for the physical, on-disk open, and one for filling out meta-data.
In short, instead of the previous interface of:

    fd = segyio.SegyFile(...)

it's now necessary to do:

    fd = _segyio.segyiofd(...)
    fd.segyopen()

    fd = _segyio.segyiofd(...)
    fd.segymake(...)

for opening and making files, respectively. This does not leak into user
code, and is taken care of inside open and create.

This properly splits the open/create routines into separate functions,
instead of interleaving them with conditionals. This is a readability
win, at the cost of some minor duplication.