Support LSB files
The SEG-Y standard and revision 1 are specified to be MSB [1]. With the
release of v2, LSB data formats are introduced. Additionally, seismic
unix files can be both MSB and LSB.

There are no significant changes in segyio's interface - segyio presents
a reality where all files are always MSB, even if the underlying file
happens to be LSB. That means segyio will translate LSB trace headers to
MSB in segy_traceheader() etc.

To tell segyio that the underlying file is LSB, call
segy_set_format(SEGY_LSB).

The test suite has gotten a new --test-lsb option that runs normal tests
on files with lsb.

[1] most-significant-byte first, or 'big-endian'