Call write, not read, in mmap write-trace
The wrong function (!) was called here, meaning memeory mapped writes
never really worked as expected. This was not caught by a test, because
the tests for this function use new files, and memory freshly-created
files has never been supported.

The write-sub-trace tests have been amended to also work with
memory-mapping. The new strategy for these tests to make a test-specific
copy of the small.sgy file, and open it in read-write mode, instead of
creating a new truncated test-specific file as before.