Initial experimental C++ interface
segyio needs a more ergonomic and simpler interface for use in
high-performance C++ applications. For now, this feature is experimental
and will only be installed on systems when the -DEXPERIMENTAL=ON flag is
passed to cmake, and is not guaranteed to be stable between even two
commits.

The interface is designed to be powerful for *use* - most programs don't
really care about SEG-Y details at all, they just want to read some
seismics. Some applications care deeply about some details (such as the
presence of fast volume/cube access). segyio-C++ was designed to support
both these cases, and to be able to statically describe required and
useful features of file handles.

The implementation should be suited for experimentation, but lacks both
maintainer documentation, user documentation, useful aliases, and a
guide for implementing new traits.