Replace ntohs/ntohl with be16toh/be32toh
An unconditional bswap16/32 is now needed, and cmake can reliably check
for host endinaness, so dependency on the htons family of functions can
be killed. This removes the dependency on ws2_32, making Windows linking
slightly easier.

On linux, there are functions for this readily availble, but there are
fewer portable alternatives. The function itself is simple enough, so
simply define it. The names be16toh/be32toh was chosen to align with
endian.h [1] or projects like [2], which can hopefully provide the
implementation in the future.

[1] http://man7.org/linux/man-pages/man3/endian.3.html
[2] https://gist.github.com/michaeljclark/3b4fd912f6fa8bb598b3