HeaderLine writes each trace just once
HeaderLine called Header with the wrong range, resulting in way to many
write operations when writing lines along the sorting  direction. I. e.
the same headers was written muliple times, tanking the performance.
Correcting the range reduced the function call complexity from  n^2
to n.

Example: Writing headers in iline mode on a iline sorted file:

1:        f.header.iline[n] = _

HeaderLine then called Header with:

2:       self.header[n0, :, stride] = _

where n0 is the first trace of iline n. In this case stride = 1, hence
we are writing all header from iline n to the end of the file. All
headers after line n is wronly assigned. However, in cases where the
entire file is updated with (1), the last assignment to each header
would still be correct, thus test has not picked up on this issue.