Add tests for sample-interval
Add tests for sample-interval that cover the behaviour when the bin
and/or trace fields are negative. The new functionality is to not select
the fallback when one of these is negative, but instead opt for the
positive one.

SEG-Y has no reasonable notion of null-values and defer it to the
implementation or even individual files to add semantics to particular
values. Since a negative sample interval is largely non-sensical,
interpret it as "missing". Users can still query the raw header value
and interpret it differently if there are some peculiar semantics.