Raise TypeError on f.trace[None]
Do the unpacking slice.indices when it is suspected the argument is a
range outside of the generator definition, and redirect the
AttributeError to a TypeError (which it really is). Otherwise the error
would first show up when the generator is accessed, instead of at the
actual site of error.