bswap16 short header words in field_forall
bswap32 cannot be indiscriminatly applied to the int32 temporary, as it
essentially is a byte-reorder and left-shift when applies to an int16.

The symptom is that certain header fields extracted this way show up in
the absurd range (some 65 million), when normal values are in the few
thousands, such as dt/sample-interval.

Addresses #368

https://github.com/equinor/segyio/issues/368