Clarify open two-file docstring example
The docstring for open.py includes an example of copying
a file from a source to a destination.

In the previous version, this example uses the same filename
for the 'from' as for the 'to' open statements.

This could be confusing to users because, in practice,
a non-trivial use of this would need different filenames for the
'from' and 'to' open statements.

This commit gives two different filenames for 'from' and 'to'.