Replace setuptools with skbuild
Replace setuptools with scikit-build [1]

Scikit-build is a nice alternatives to raw setuptools when building
native-extensions, and a perfect fit for a project with segyio's
structure. There is a minor breaking change for some build routines -
the python setup script is moved the the python/ directory. This does
not come without benefits:

1. The python project can now be completely oblivious, save the root
   cmake file, to the fact that it's a part of a larger repository, and
   behaves like most python projects
2. cmake is used to build the extension, which means compiler options
   and configs can be set in a more ergonomic manner
3. Because cmake and scikit now manages the compiler setup and binary
   compatibility, compiler support might just be extended, even for
   Windows. segyio no longer has to maintain or deal with these details,
   and can leave it to scikit-build

[1] https://github.com/scikit-build/scikit-build