Point out sparse file writes in common issues
The issue of slow header writes pop up quite often, and it's likely that
the culprit is actually a slow file system for sparse writes. It is
certainly useful to point this out in the common issues section in the
readme, as well as briefly mention sparse writes in the docs for create.