Use multibuild as clone, not submodule
Multibuild, some useful automation for wheels on travis, was always
fetched as a submodule. This ensured stability some stability for the
testing environment, as it was assumed that multibuild might at times
be volatile. Experience from other projects have demonstrated this is
not the case, and just cloning it is fine.

This means downstream packaging no longer have to deal with the empty
multibuild directory, and upstream fixes that respond to the travis
environment are picked up faster.