Add wide binary header word replacements from rev2
SEG-Y rev2 introduces a handful of replacement header words with larger
size for high trace and sample counts. Add them to the binfield, words,
and size enums so the fields are easily accessible to users.

Please note that the split of the revision word into a minor/major is
fixed with this.