Add support for int24 formats
Flip the switch on the sleeping int24 support. This primarily means
fixing a bug [1] in the byteswap code, and returning 3 as elem-size on
format-code 7 and 15.

Most of the code in this patch is dedicated to the tests, which like all
the other format tests are based on the f3 data. Format7/15 files were
manually converted into int24 with a one-off C++ program with this
operation for every trace:

    for (int i = 0; i < 75; ++i) {
        char buf[3];
        buf[0] = *(buffer.data() + i * 2 + 0);
        buf[1] = *(buffer.data() + i * 2 + 1);
        buf[2] = 0;
        std::memcpy(buffer3.data() + i * 3, buf, 3);
    }

This snippet is for little-endian, with the order of bytes reversed for
big-endian.

I manually verified that tests failed on byte-swap without the 24 -> 3
change in bswap24vec, but other than that testing is not very thorough
as I've yet to see an int24 file in the wild.

Please note that since there is no int24 support in numpy, Python still
does not support int24 and will fail when opening the file.

[1] https://github.com/equinor/segyio/issues/469