Bump version to 1.9.9
We already have an tag for version 1.9.8, but the wheels never got
deployed, and we forgot to bump the version in setup.cfg and update the
changelog.

We could still bump to 1.9.8 and push the wheels to pypi at this time,
but that would mean that the release on pypi is out-of-sync with the git
tag and the source release on github. A better approach seams to be to
just acknowledge that no wheels where ever published for 1.9.8 and bump
to 1.9.9.